pyproject.toml
[tool.poetry]
name = "kuki"
version = "0.1.0"
description = ""
authors = ["Neil Van Alino <nvq.alino@gmail.com>"]
license = "GNU GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pywebview = {extras = ["qt"], version = "^5.3.2"}
streamlit = "^1.40.0"
pyinstaller = "^6.11.0"
python-dotenv = "^1.0.1"
faster-whisper = "^1.0.3"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
mypy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
